server4: Fix peer address logic
The RFC has this to say about how to select the address to respond to:

>  If the 'giaddr' field in a DHCP message from a client is non-zero,
>  the server sends any return messages to the 'DHCP server' port on the
>  BOOTP relay agent whose address appears in 'giaddr'. If the 'giaddr'
>  field is zero and the 'ciaddr' field is nonzero, then the server
>  unicasts DHCPOFFER and DHCPACK messages to the address in 'ciaddr'.
>  If 'giaddr' is zero and 'ciaddr' is zero, and the broadcast bit is
>  set, then the server broadcasts DHCPOFFER and DHCPACK messages to
>  0xffffffff. If the broadcast bit is not set and 'giaddr' is zero and
>  'ciaddr' is zero, then the server unicasts DHCPOFFER and DHCPACK
>  messages to the client's hardware address and 'yiaddr' address.  In
>  all cases, when 'giaddr' is zero, the server broadcasts any DHCPNAK
>  messages to 0xffffffff.

We can implement almost all of this. Unfortunately the bit about non-broadcast
responses ("If the broadcast bit is not set and 'giaddr' is zero and
'ciaddr' is zero, then the server unicasts DHCPOFFER and DHCPACK
messages to the client's hardware address and 'yiaddr' address")
is impossible to implement with the current API offered by go and the
dhcp library (it requires raw sockets and a way to pass L2 information
in the peer address).

The RFC also says this:

> If unicasting is not possible, the message MAY be sent as an IP
> broadcast using an IP broadcast address (preferably 0xffffffff) as the
> IP destination address and the link-layer broadcast address as the
> link-layer destination address.

Which we'll do for the time being

Signed-off-by: Anatole Denis <anatole@unverle.fr>