config: Expand unspecified ll-mcast to all interfaces
Specifying a link-local multicast address to listen on without an
interface is rejected. Instead, list all available interfaces where it
makes sense ("MULTICAST" flag) and listen on all of them, which is the
least-surprising behaviour.

This will be needed to have sane default values since the RFC specifies
a default link-local multicast address to listen on

Document a surprising behaviour when interfaces appear and disappear
during runtime. Note it is *not* a regression as the previous behaviour
was worse (join the mcast group on only one interface with a default
route)

Signed-off-by: Anatole Denis <anatole@unverle.fr>