plugins/netmask: Don't panic on invalid netmask
It looks like the error handling for this plugin is borked. I noticed
this while reading this code during my attempt to implement a DHCP
server. This also clarifies the error messages, since they were either
innaccurate, or duplicated which would make it impossible to tell which
code path the error actually came from.