config: Re-add interface keyword to keep compatibility
This re-adds an `interface` keyword in the configuration to keep
compatibility with existing configuration files. The behaviour is
emulated in terms of the `listen` keyword.
Both cannot be used together, but since listen wasn't an accepted
keyword previously, users are expected to remove the old one when they
add the new one

Signed-off-by: Anatole Denis <anatole@unverle.fr>